체력과 공격력의 자료형을 정수형으로 변경. 스킬의 데미지를 추가 데미지로 변경. 무기 데미지에 스킬 데미지를 합산하여 타격에 적용하도록 변경

+++ Assets/2.Script/Core/Combat System/Interface/ICombatable.cs
@@ -6,5 +6,5 @@ public interface ICombatable
     /// 체력을 증감하는 함수
     /// </summary>
     /// <param name="damage">증감 수치</param>
-    public void TakeHit(float damage, BaseObject hitter);
+    public void TakeHit(int damage, BaseObject hitter);
 }
